+++ src/frontend/src/pages/MainPage/pages/deploymentsPage/components/step-attach-flows-connection-panel.tsx
@@ -1,11 +1,10 @@
-import { memo, useMemo, useState } from "react";
-import ForwardedIconComponent from "@/components/common/genericIconComponent";
+import { memo } from "react";
 import InputComponent from "@/components/core/parameterRenderComponent/components/inputComponent";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { cn } from "@/utils/utils";
 import type { ConnectionItem, EnvVarEntry } from "../types";
-import { CheckboxSelectItem } from "./radio-select-item";
+import { ConnectionSearchList } from "./connection-search-list";
 
 export type ConnectionTab = "available" | "create";
 
@@ -48,21 +47,6 @@ export const ConnectionPanel = memo(function ConnectionPanel({
   onCreateConnection: () => void;
   isDuplicateName?: boolean;
 }) {
-  const [searchQuery, setSearchQuery] = useState("");
-  const filteredConnections = useMemo(() => {
-    // Sort newly created connections to the top
-    const sorted = [...connections].sort((a, b) => {
-      if (a.isNew && !b.isNew) return -1;
-      if (!a.isNew && b.isNew) return 1;
-      return 0;
-    });
-    if (!searchQuery.trim()) return sorted;
-    const q = searchQuery.toLowerCase();
-    return sorted.filter(
-      (c) => c.name.toLowerCase().includes(q) || c.id.toLowerCase().includes(q),
-    );
-  }, [connections, searchQuery]);
-
   return (
     <>
       <div className="border-b border-border p-4 text-sm text-muted-foreground">
@@ -93,65 +77,18 @@ export const ConnectionPanel = memo(function ConnectionPanel({
         </div>
 
         {/* Tab content */}
-        <div className="mt-4 flex-1 overflow-x-hidden overflow-y-auto">
+        <div
+          className="mt-4 flex-1 overflow-x-hidden overflow-y-auto"
+          key={connectionTab}
+        >
           {connectionTab === "available" ? (
             <div className="min-w-0 space-y-3">
-              {connections.length === 0 ? (
-                <div className="flex flex-col items-center justify-center gap-3 py-12 text-center">
-                  <ForwardedIconComponent
-                    name="PlugZap"
-                    className="h-8 w-8 text-muted-foreground/50"
-                  />
-                  <div>
-                    <p className="text-sm font-medium text-muted-foreground">
-                      No connections yet
-                    </p>
-                    <p className="mt-0.5 text-xs text-muted-foreground/70">
-                      Create a connection to attach credentials to this flow.
-                    </p>
-                  </div>
-                  <button
-                    type="button"
-                    onClick={() => onTabChange("create")}
-                    className="text-xs font-medium text-primary hover:underline"
-                  >
-                    Create your first connection
-                  </button>
-                </div>
-              ) : (
-                <>
-                  <div className="min-w-0">
-                    <Input
-                      icon="Search"
-                      placeholder="Search connections..."
-                      className="bg-muted"
-                      value={searchQuery}
-                      onChange={(e) => setSearchQuery(e.target.value)}
-                    />
-                  </div>
-                  {filteredConnections.length === 0 ? (
-                    <p className="py-6 text-center text-sm text-muted-foreground">
-                      No connections match &ldquo;{searchQuery}&rdquo;
-                    </p>
-                  ) : (
-                    filteredConnections.map((conn) => (
-                      <CheckboxSelectItem
-                        key={conn.id}
-                        value={conn.id}
-                        checked={selectedConnections.has(conn.id)}
-                        onChange={() => onToggleConnection(conn.id)}
-                        data-testid={`connection-item-${conn.id}`}
-                      >
-                        <div className="min-w-0 flex-1">
-                          <span className="block truncate text-sm font-medium leading-tight">
-                            {conn.name}
-                          </span>
-                        </div>
-                      </CheckboxSelectItem>
-                    ))
-                  )}
-                </>
-              )}
+              <ConnectionSearchList
+                connections={connections}
+                selectedConnections={selectedConnections}
+                onToggleConnection={onToggleConnection}
+                onSwitchToCreate={() => onTabChange("create")}
+              />
             </div>
           ) : (
             <div className="flex flex-col gap-4">
